requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LLM and AI
openai==1.3.7
langchain==0.1.0
langchain-community==0.0.10
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
ollama==0.1.6

# RAG and Vector Stores
pinecone-client==3.0.0
chromadb==0.4.18
faiss-cpu==1.7.4
weaviate-client==3.25.3

# Infrastructure Tools
python-terraform==0.10.1
kubernetes==28.1.0
ansible-runner==2.3.4
docker==6.1.3

# Security and Validation
checkov==3.1.0
bandit==1.7.5
safety==2.3.5

# Database and Storage
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1
psycopg2-binary==2.9.9

# API and Web
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Git Integration
gitpython==3.1.40

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
pyyaml==6.0.1
jinja2==3.1.2
click==8.1.7
rich==13.7.0
tenacity==8.2.3
